algo work

+++ utils/ft_atoi.c
@@ -19,7 +19,7 @@
 // 	return (0);
 // }
 
-int	ft_atoi(char *str)
+long	ft_atoi(char *str)
 {
 	int			i;
 	int			signe;
